pyproject.toml
[project]
name = "hedonometer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==4.2",
    "django-tastypie==0.14.6",
    "jinja2>=3.0",
]
